Advanced commit: ========================
Users can now be added via swagger API.
At the moment passwords are being encrypted and stored as such in the database.
Added the 3 DTOs along with their respective functionality.
Now have to add the view, update, and delete parts for the users and once that's finished advance to the images part of the project.

+++ PhotoMemoriesDomain/src/main/java/com/photomemories/domain/persistence/Shared.java
@@ -15,7 +15,7 @@ public class Shared implements Serializable {
     //TODO considering to make use of Integer instead of Long
     @Id
     @Column(name = "SHARED_ID")
-    private Long SharedId;
+    private Integer SharedId;
 
     @Column(name = "SH_SHARED_DATE")
     private Long SharedBy;
@@ -39,18 +39,18 @@ public class Shared implements Serializable {
         this.PhotoId = photoId;
     }
 
-    public Shared(Long sharedId, Long sharedBy, User userId, Photo photoId) {
+    public Shared(Integer sharedId, Long sharedBy, User userId, Photo photoId) {
         this.SharedId = sharedId;
         this.SharedBy = sharedBy;
         this.UserId = userId;
         this.PhotoId = photoId;
     }
 
-    public Long getSharedId() {
+    public Integer getSharedId() {
         return SharedId;
     }
 
-    public void setSharedId(Long sharedId) {
+    public void setSharedId(Integer sharedId) {
         SharedId = sharedId;
     }
